+++ osu.Game/Graphics/UserInterfaceV2/LabelledDrawable.cs
@@ -5,6 +5,7 @@ using osu.Framework.Allocation;
 using osu.Framework.Graphics;
 using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Shapes;
+using osu.Framework.Extensions.Color4Extensions;
 using osu.Game.Graphics.Containers;
 using osuTK;
 
@@ -42,7 +43,7 @@ namespace osu.Game.Graphics.UserInterfaceV2
                 new Box
                 {
                     RelativeSizeAxes = Axes.Both,
-                    Colour = OsuColour.FromHex("1c2125"),
+                    Colour = Color4Extensions.FromHex("1c2125"),
                 },
                 new FillFlowContainer
                 {