+++ osu.Game/Overlays/Direct/Header.cs
@@ -8,12 +8,13 @@ using osu.Framework.Graphics.Sprites;
 using osu.Game.Graphics;
 using osu.Game.Graphics.Sprites;
 using osu.Game.Overlays.SearchableList;
+using osu.Framework.Extensions.Color4Extensions;
 
 namespace osu.Game.Overlays.Direct
 {
     public class Header : SearchableListHeader<DirectTab>
     {
-        protected override Color4 BackgroundColour => OsuColour.FromHex(@"252f3a");
+        protected override Color4 BackgroundColour => Color4Extensions.FromHex(@"252f3a");
 
         protected override DirectTab DefaultTab => DirectTab.Search;
         protected override Drawable CreateHeaderText() => new OsuSpriteText { Text = @"osu!direct", Font = OsuFont.GetFont(size: 25) };