+++ osu.Game/Overlays/Social/Header.cs
@@ -10,6 +10,7 @@ using osu.Game.Graphics;
 using osu.Framework.Allocation;
 using System.ComponentModel;
 using osu.Framework.Graphics.Sprites;
+using osu.Framework.Extensions.Color4Extensions;
 
 namespace osu.Game.Overlays.Social
 {
@@ -17,7 +18,7 @@ namespace osu.Game.Overlays.Social
     {
         private OsuSpriteText browser;
 
-        protected override Color4 BackgroundColour => OsuColour.FromHex(@"38202e");
+        protected override Color4 BackgroundColour => Color4Extensions.FromHex(@"38202e");
 
         protected override SocialTab DefaultTab => SocialTab.AllPlayers;
         protected override IconUsage Icon => FontAwesome.Solid.Users;